+++ spec/castkit/data_object_spec.rb
@@ -17,6 +17,10 @@ RSpec.describe Castkit::DataObject do
 
   let(:valid_input) { { name: "Tester", age: 30 } }
 
+  before do
+    allow(Castkit).to receive(:warning)
+  end
+
   describe ".cast" do
     it "returns the object if it's already an instance" do
       instance = subclass.new(valid_input)
@@ -49,10 +53,7 @@ RSpec.describe Castkit::DataObject do
   describe ".dump" do
     it "delegates to_json on the object" do
       instance = subclass.new(valid_input)
-      expect(JSON.parse(subclass.dump(instance))).to eq({
-                                                          "name" => "Tester",
-                                                          "age" => 30
-                                                        })
+      expect(JSON.parse(subclass.dump(instance))).to eq({ "name" => "Tester", "age" => 30 })
     end
   end
 
@@ -78,13 +79,55 @@ RSpec.describe Castkit::DataObject do
       end.to raise_error(Castkit::DataObjectError, /Unknown attribute/)
     end
 
+    it "raises on unknown keys in strict mode when allow_unknown is also false" do
+      subclass.strict true
+      subclass.allow_unknown false
+
+      expect do
+        subclass.new(valid_input.merge(extra: 1))
+      end.to raise_error(Castkit::DataObjectError, /Unknown attribute/)
+    end
+
     it "warns on unknown keys in warn mode" do
       subclass.strict false
       subclass.warn_on_unknown true
 
-      expect do
-        subclass.new(valid_input.merge(foo: 1))
-      end.to output(/Unknown attribute.*foo/).to_stderr
+      subclass.new(valid_input.merge(foo: 1))
+      expect(Castkit).to have_received(:warning).with(/Unknown attribute.*foo/)
+    end
+
+    it "warns when strict is set to true and allow_unknown is set to true" do
+      subclass.strict true
+      subclass.allow_unknown true
+      subclass.new(valid_input)
+
+      expect(Castkit).to have_received(:warning).with(/`strict` and `allow_unknown` are enabled/)
+    end
+
+    it "overrides strict when allow_unknown is true" do
+      subclass.strict true
+      subclass.allow_unknown true
+
+      expect { subclass.new(valid_input.merge(extra: 1)) }.not_to raise_error
+    end
+  end
+
+  describe "#__raw" do
+    it "returns the raw data on instantiation" do
+      subclass.root :person
+      input = { person: { name: "Wrapped", age: 20 } }
+
+      instance = subclass.new(input)
+      expect(instance.__raw).to eq(input)
+    end
+  end
+
+  describe "#unknown_attributes" do
+    it "returns undefined (unknown) attributes" do
+      subclass.allow_unknown true
+
+      instance = subclass.new(valid_input.merge(extra: 1))
+      expect(instance.unknown_attributes).to eq({ extra: 1 })
     end
   end
 
@@ -93,6 +136,13 @@ RSpec.describe Castkit::DataObject do
       instance = subclass.new(valid_input)
       expect(instance.to_hash).to eq({ name: "Tester", age: 30 })
       expect(instance.serialize).to eq(instance.to_hash)
+    end
+
+    it "returns the serialized hash with unknown attributes" do
+      subclass.allow_unknown true
+
+      instance = subclass.new(valid_input.merge(extra: 1))
+      expect(instance.to_hash).to eq({ name: "Tester", age: 30, extra: 1 })
     end
   end